requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2

